requirements.txt
pandas==2.3.2
numpy==2.2.1
scikit-learn==1.5.2
joblib==1.4.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-storage==2.18.2
